7/22
This is the updated export script, with proper column and row finding information. This version is currently functioning, but not behaving properly. When rendering images, instead of searching across the pole and giving us a better hemispherical arrangement as in the first frame of Chimera, the particles seem to be searching for neighbors farther in towards the PNS, creating a large beaming artifact with two spheres located symmetrically across the vertical origin of the progenitor.